warning
git-svn-id: http://svn.kcn.cn/repos/kbs/trunk@9444 de4889a5-8ad4-4bf3-bd87-57997c21d515

+++ kbs_bbs/src/announce.c
@@ -356,7 +356,7 @@ MENU *pm;
     } else
         strncpy(genbuf, pm->mtitle, MAXPATH);
     if (strlen(genbuf) <= 80)
-        sprintf(buf, "%*s", (80 - strlen(genbuf)) / 2, "");
+        sprintf(buf, "%*s", (int)( (80 - strlen(genbuf)) / 2 ), "");
     else
         strcpy(buf, "");
     prints("\033[44m%s%s%s\033[m\n", buf, genbuf, buf);
